Why would I do that?

+++ BrainfuckInterpreter/src/main.cpp
@@ -17,7 +17,7 @@
 
 #include "BfInterpreter.hpp"
 
-int main(size_t argc, char** argv)
+int main(int argc, char** argv)
 {
     try
     {
